refact: move Spanned.extensions to 'spanned' package
